requirements.txt
flask
flask_sqlalchemy
flask_cors
python-dotenv
pyjwt
